+++ packages/addfox/CHANGELOG.md
@@ -1,5 +1,14 @@
 # addfox
 
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @addfox/core@0.2.8
+  - @addfox/cli@0.2.9
+  - @addfox/rsbuild-plugin-extension-monitor@0.2.8
+
 ## 0.2.8
 
 ### Patch Changes