+++ packages/plugins/rsbuild-plugin-extension-entry/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @addfox/rsbuild-plugin-extension-entry
 
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @addfox/core@0.2.8
+
 ## 0.2.8
 
 ### Patch Changes